+++ packages/react-frontend/src/components/Carousel.css
