+++ packages/react-frontend/src/components/MovieCard.css
