+++ packages/react-frontend/src/components/MovieCard.js
