+++ packages/react-frontend/src/components/MovieList.js
