+++ packages/react-frontend/src/pages/LoginPage.css
@@ -1,7 +1,7 @@
 .login {
     position: relative;
     height: 100%;
-    background: url("./assets/signup_background.jpg") center no-repeat;
+    background: url("../assets/signup_background.jpg") center no-repeat;
     background-size: cover;
     font-family: sans-serif;
 }
