+++ packages/react-frontend/src/pages/MovieReviewPage.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Header from "./Header";
+import Header from "../components/Header";
 import "./MovieReviewPage.css";
 
 function MovieReviewPage(props) {
